Moved external_interfaces/common over to cpp implementation files.
